pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pipecat-deepdub-tts"
version = "0.1.0"
description = "Official Deepdub AI TTS integration for Pipecat"
readme = "README.md"
license = "MIT"
authors = [
    {name = "Deepdub AI", email = "info@deepdub.ai"},
]
requires-python = ">=3.10,<3.13"
keywords = ["deepdub", "pipecat", "text-to-speech", "tts", "voice-ai", "conversational-ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "deepdub>=0.1.20",
    "loguru>=0.7.3",
    "pipecat-ai>=0.0.97,<0.1.0",
    "websockets>=15.0.1,<16.0",
]

[project.urls]
Repository = "https://github.com/deepdub-ai/pipecat-deepdub-tts"
Issues = "https://github.com/deepdub-ai/pipecat-deepdub-tts/issues"

[dependency-groups]
dev = [
    "fastapi>=0.128.8",
    "pipecat-ai[assemblyai,openai,silero]>=0.0.97,<0.1.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
